[DD/MM/YYYY] * UPDATE *  - Savepoint -> rest point  - NPC Quest  - Quest System UI  - Quest Window * FIX * - - * REMOVE * - Dialog controller - * TODO * - add some sound, enemy attack, level up, background music...

+++ Assets/_Script/BarSystemController.cs
@@ -45,9 +45,9 @@ public class BarSystemController : MonoBehaviour
     private void UpdateFillAmount_Text(Image image, TextMeshProUGUI text , float min, float max) // Fill Amount and text UI
     {
         image.fillAmount = min / max;
-        text.text = min.ToString() + "/" + max.ToString();
+        text.text = ((int)min).ToString() + "/" + max.ToString();
     }
-    private void UpdateFillAmount(Image image, float min, float max) // Full amount only
+    private void UpdateFillAmount(Image image, float min, float max) // Fill amount only
     {
         image.fillAmount = min / max;
     }
